+++ src/components/sections/brands-section.tsx
@@ -60,6 +60,7 @@ export async function BrandsSection() {
                         width={24}
                         height={24}
                         className="h-6 w-6 object-contain"
+                        unoptimized
                       />
                       {brand.name}
                     </Badge>